DRTII-1305 Upgrade MUI & MUI X. Migrate & clean up
