+++ react/package.json
@@ -6,13 +6,12 @@
     "@date-io/moment": "^2.11.0",
     "@emotion/react": "^11.4.1",
     "@emotion/styled": "^11.3.0",
-    "@mui/icons-material": "^5.0.0",
-    "@mui/lab": "^5.0.0-alpha.47",
-    "@mui/material": "^5.9.2",
-    "@mui/styled-engine": "^5.0.0",
-    "@mui/x-data-grid": "5.17.3",
-    "@mui/x-data-grid-generator": "5.17.3",
-    "@mui/x-date-pickers": "^5.0.0-beta.3",
+    "@mui/icons-material": "^5.14.16",
+    "@mui/lab": "^5.0.0-alpha.152",
+    "@mui/material": "^5.14.17",
+    "@mui/styled-engine": "^5.14.17",
+    "@mui/x-data-grid": "^6.18.0",
+    "@mui/x-date-pickers": "^6.18.0",
     "@reduxjs/toolkit": "^1.9.7",
     "@testing-library/react": "^11.1.2",
     "@testing-library/user-event": "^7.1.2",
