+++ react/src/App.tsx
@@ -1,9 +1,9 @@
 import React, {useEffect} from 'react';
 import './App.css';
 import {Home} from './components/Home';
-import Alerts from './components/Alerts/Alerts';
-import AccessRequests from './components/UserManagement/AccessRequests';
-import UserTracking from './components/users/UserTracking';
+import Alerts from './components/alerts/Alerts';
+import AccessRequests from './components/accessrequests/AccessRequests';
+import UsersList from './components/users/UsersList';
 import {Route, Routes} from "react-router-dom";
 import Loading from "./components/Loading";
 import Navigation from "./components/Navigation";
@@ -13,10 +13,13 @@ import {styled} from "@mui/material/styles";
 import {RegionPage} from "./components/RegionPage";
 import axios from "axios";
 import ApiClient from "./services/ApiClient";
-import UploadForm from "./components/featureGuide/FeatureGuideUploadFile";
-import {DropInLanding} from "./components/dropin/DropInLanding";
+import UploadForm from "./components/featureguides/FeatureGuideUploadFile";
 import {HealthCheckEditor} from "./components/HealthCheckPausesEditor";
 import {useUser} from "./store/userSlice";
+import {DropInsList} from "./components/dropins/DropInsList";
+import {CreateDropIn} from "./components/dropins/CreateDropIn";
+import {EditDropIn} from "./components/dropins/EditDropIn";
+import {RegisteredUsers} from "./components/dropins/RegisteredUsers";
 
 const StyledDiv = styled('div')(() => ({
   textAlign: 'center',
@@ -80,11 +83,18 @@ export const App = () => {
         <Routes>
           <Route path="/" element={<Home config={config.values} user={user.profile}/>}/>
           <Route path="/access-requests" element={<AccessRequests/>}/>
-          <Route path="/users" element={<UserTracking/>}/>
+          <Route path="/users" element={<UsersList/>}/>
           <Route path="/alerts" element={<Alerts regions={config.values.portsByRegion} user={user.profile}/>}/>
           <Route path="/region/:regionName" element={<RegionPage user={user.profile} config={config.values}/>}/>
           <Route path="/feature-guide-upload" element={<UploadForm/>}/>
-          <Route path="/drop-ins/list" element={<DropInLanding/>}/>
+          <Route path="/drop-ins">
+            <Route path="" element={<DropInsList/>}/>
+            <Route path="list/crud/:operations" element={<DropInsList/>}/>
+            <Route path="list/:listAll?" element={<DropInsList/>}/>
+            <Route path="new" element={<CreateDropIn/>}/>
+            <Route path="edit/:dropInId" element={<EditDropIn/>}/>
+            <Route path="list/registeredUsers/:dropInId" element={<RegisteredUsers/>}/>
+          </Route>
           <Route path="/health-checks" element={<HealthCheckEditor/>}/>
         </Routes>
       </StyledContainer>
