+++ react/src/components/HealthCheckPausesEditor.tsx
@@ -1,16 +1,5 @@
 import React, {useState} from 'react';
-import {
-  Box,
-  Button,
-  Dialog,
-  DialogActions,
-  DialogContent,
-  DialogTitle,
-  Grid,
-  Snackbar,
-  Stack,
-  TextField
-} from "@mui/material";
+import {Box, Button, Dialog, DialogActions, DialogContent, DialogTitle, Grid, Snackbar, Stack} from "@mui/material";
 import {Cancel, Delete, Save} from "@mui/icons-material";
 import {ScheduledHealthCheckPause} from "./healthcheckpauseseditor/model";
 import moment from "moment-timezone";
@@ -22,7 +11,7 @@ import {AdapterMoment} from "@mui/x-date-pickers/AdapterMoment";
 import {LocalizationProvider} from '@mui/x-date-pickers/LocalizationProvider';
 import {Moment} from "moment";
 import {DataGrid} from "@mui/x-data-grid";
-
+import {Alert} from "@mui/lab";
 
 type ConfirmOpen = {
   kind: 'open'
@@ -42,19 +31,21 @@ export const HealthCheckEditor = () => {
   const [newPause, setNewPause] = useState<ScheduledHealthCheckPause | null>(null)
   const [confirm, setConfirm] = useState<Confirm>({kind: 'closed'})
   const [snackbarMessage, setSnackbarMessage] = useState<string | null>(null)
+  const [newPauseIsValid, setNewPauseIsValid] = useState<boolean>(true)
+
+  moment.locale('en-gb')
 
   const columns = [
     {
-      field: 'startsAt',
-      headerName: 'From',
-      valueGetter: (params: any) => moment(params.row.startsAt).format("HH:mm, Do MMM YYYY"),
-      width: 200
-    },
-    {
-      field: 'endsAt',
-      headerName: 'To',
-      valueGetter: (params: any) => moment(params.row.endsAt).format("HH:mm, Do MMM YYYY"),
-      width: 200
+      field: 'pause',
+      headerName: 'Pause',
+      valueGetter: (params: any) => {
+        const sameStartDate = moment(params.row.startsAt).isSame(moment(params.row.endsAt), 'day')
+        return sameStartDate ?
+          `${moment(params.row.startsAt).format("HH:mm")} to ${moment(params.row.endsAt).format("HH:mm, Do MMM YYYY")}` :
+          `${moment(params.row.startsAt).format("HH:mm, Do MMM YYYY")} to ${moment(params.row.endsAt).format("HH:mm, Do MMM YYYY")}`
+      },
+      width: 400
     },
     {
       field: 'createdAt',
@@ -78,25 +69,40 @@ export const HealthCheckEditor = () => {
   ]
 
 
-  const today: () => Moment = () => {
-    return moment()
-  }
+  const today: () => Moment = () => moment()
 
   function addNewPause() {
+    const now = today()
+    const monthOfTheYear = now.month() + 1
+    const str = `${now.year()}-${monthOfTheYear.toString().padStart(2, '0')}-${now.date().toString().padStart(2, '0')}T${now.hour() + 1}:00:00.000Z`
+    const nextHour = moment(str)
+    console.log(`str: ${str}, nextHour: ${nextHour.format()}`)
     setNewPause({
-      startsAt: today().valueOf(),
-      endsAt: today().valueOf(),
+      startsAt: nextHour.valueOf(),
+      endsAt: nextHour.add(1, 'hour').valueOf(),
       ports: [],
-      createdAt: today().valueOf(),
+      createdAt: now.valueOf(),
     })
   }
 
   const setStart: (e: Moment | null) => void = e => {
-    (newPause && e) && setNewPause({...newPause, startsAt: e.valueOf()})
+    if (newPause && e) {
+      const pause = {...newPause, startsAt: e.valueOf()}
+      setNewPause(pause)
+      validate(pause)
+    }
   }
 
   const setEnd: (e: Moment | null) => void = e => {
-    (newPause && e) && setNewPause({...newPause, endsAt: e.valueOf()})
+    if (newPause && e) {
+      const pause = {...newPause, endsAt: e.valueOf()}
+      setNewPause(pause)
+      validate(pause)
+    }
+  }
+
+  const validate = (pause: ScheduledHealthCheckPause) => {
+    setNewPauseIsValid(pause.startsAt < pause.endsAt)
   }
 
   const saveNewPause = (pause: ScheduledHealthCheckPause) => {
@@ -136,11 +142,12 @@ export const HealthCheckEditor = () => {
     deletePause(from, to)
   }
 
-  const rows = healthCheckPauses.sort((a, b) => -1 * (a.startsAt.valueOf() - b.startsAt.valueOf()))
+  const rows = [...healthCheckPauses]
+    .sort((a, b) => -1 * (b.startsAt.valueOf() - a.startsAt.valueOf()))
 
   return (
     <Stack sx={{my: 2, gap: 2}}>
-      <LocalizationProvider dateAdapter={AdapterMoment}>
+      <LocalizationProvider dateAdapter={AdapterMoment} adapterLocale={'en-gb'}>
         <Snackbar
           anchorOrigin={{vertical: 'top', horizontal: 'center'}}
           open={!!snackbarMessage}
@@ -162,22 +169,28 @@ export const HealthCheckEditor = () => {
                   </DialogTitle>
                   <DialogContent sx={{minWidth: 380}}>
                       <Stack sx={{my: 2, gap: 2}}>
-                          <DateTimePicker renderInput={(params) => <TextField {...params}/>}
-                                          value={moment(newPause.startsAt)}
-                                          label={'From'}
-                                          onChange={setStart}/>
-                          <DateTimePicker renderInput={(params) => <TextField {...params}/>}
-                                          value={moment(newPause.endsAt)}
-                                          label={'To'}
-                                          onChange={setEnd}/>
+                        {!newPauseIsValid && <Alert severity="error">The start time must be before the end time</Alert>}
+                          <DateTimePicker
+                              slotProps={{textField: {variant: 'outlined'}}}
+                              value={moment(newPause.startsAt)}
+                              label={'From'}
+                              onChange={setStart}/>
+                          <DateTimePicker
+                              slotProps={{textField: {variant: 'outlined'}}}
+                              value={moment(newPause.endsAt)}
+                              label={'To'}
+                              onChange={setEnd}
+                          />
                       </Stack>
                   </DialogContent>
                   <DialogActions>
-                      <Button color="primary" variant="outlined" size="medium" onClick={() => setNewPause(null)}>
+                      <Button color="primary" variant="outlined" size="medium"
+                              onClick={() => setNewPause(null)}>
                           <Cancel/> Cancel
                       </Button>
                       <Button color="primary" variant="outlined" size="medium"
-                              onClick={() => newPause && saveNewPause(newPause)}>
+                              onClick={() => newPause && saveNewPause(newPause)}
+                              disabled={!newPauseIsValid}>
                           <Save/> Save
                       </Button>
                   </DialogActions>
@@ -201,13 +214,13 @@ export const HealthCheckEditor = () => {
             failed ?
               <Typography variant={'body1'}>Sorry, I couldn't load the existing pauses</Typography> :
               healthCheckPauses.length === 0 ?
-                <Typography variant={'body1'}>There are no existing pauses</Typography> :
+                <Typography variant={'body1'}>There are no current or upcoming pauses</Typography> :
                 <Box sx={{height: 400, width: '100%'}}>
                   <DataGrid
                     rows={rows}
                     columns={columns}
                     getRowId={(r) => r.createdAt.valueOf().toString()}
-                    disableSelectionOnClick={true}
+                    disableRowSelectionOnClick={true}
                   />
                 </Box>
           }
