+++ react/src/components/Navigation.tsx
@@ -29,9 +29,9 @@ export default function Navigation(props: IProps) {
     {label: 'Home', link: '/', roles: []},
     {label: 'Access requests', link: '/access-requests', roles: ['manage-users']},
     {label: 'Alert notices', link: '/alerts', roles: ['manage-users']},
+    {label: 'Drop-ins', link: '/drop-ins/list', roles: ['manage-users']},
     {label: 'Feature guides', link: '/feature-guide-upload', roles: ['manage-users']},
     {label: 'Health checks', link: '/health-checks', roles: ['manage-users']},
-    {label: 'Drop-ins', link: '/drop-ins/list', roles: ['manage-users']},
     {label: 'Users', link: '/users', roles: ['manage-users']},
     {label: 'Log out', link: props.logoutLink, roles: []},
   ]
