+++ react/src/components/featureguides/DeleteComponent.tsx
